+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- `kebabCase` extension
+
+### Changed
+- `validateArgument` accepts undefined value
+- Code refactor
+- Tests migration to Given-When-Then
+- Maintenance updates